backends/mlx: address index review feedback
Move MLX index regression coverage into the consolidated op tests and document why F.unfold needs PyTorch-compatible gather ordering.
